Move method, it's needed no just in Netty but all server backends
